requirements.txt
ag2[openai]
ag2[gemini]
ag2[anthropic]
arxiv >=2.2.0
agentops>=0.4.18
langgraph>=0.5.0
langchain-openai>=0.3.24
langchain-mcp-adapters>=0.1.7
langchain_community>=0.3.26
python-dotenv>=1.1.0
openai >= 1.88.0
anthropic>=0.51.0
mcp>=1.12.4
pandas>=2.3.0
pypdf2>=3.0.1
pyTigerGraph>=1.9.0
typing
uv
wikipedia
numexpr
